Adding a conditional statement
Adding a conditional statement, in case it has been installed before

+++ scripts/install-dirsearch.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
-cd /opt
-git clone https://github.com/maurosoria/dirsearch.git
-cd dirsearch
-ln -s dirsearch.py /usr/bin/dirsearch.py
+FILE=/opt/dirsearch/dirsearch.py
+if [ -f "$FILE" ]; then
+    	echo "$FILE exist"
+else
+	cd /opt
+	git clone https://github.com/maurosoria/dirsearch.git
+	cd dirsearch
+	ln -s dirsearch.py /usr/bin/dirsearch.py
+fi
